Move SettingsForm to AppTemplate.
