+++ Sandra.UI/Sandra.UI.csproj
@@ -70,6 +70,7 @@
     </Compile>
     <Compile Include="AppTemplate\Session.cs" />
     <Compile Include="AppTemplate\Session.ToolForms.cs" />
+    <Compile Include="AppTemplate\SettingsForm.cs" />
     <Compile Include="AppTemplate\SharedLocalizedlStringKeys.cs" />
     <Compile Include="AppTemplate\SharedResources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -116,13 +117,6 @@
     </Compile>
     <Compile Include="PropertyStore.cs" />
     <Compile Include="Settings.cs" />
-    <Compile Include="SettingsForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="SettingsForm.UIActions.cs">
-      <DependentUpon>SettingsForm.cs</DependentUpon>
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Include="StandardChessBoardForm.cs">
       <SubType>Form</SubType>
     </Compile>
